initial commenting implementation
schema, model, controller, and polymer components for basic
commenting functionality. refactored the uploads and files modules a bit
so they made more sense with the comments module. need to experiment
with the UX (deleting comments, an explicit save button, a way to add
comments while still allowing for copying, etc.)